diery_server/requirements.txt
Flask==2.2.2
gunicorn
Werkzeug==2.2.2
openai
python-dotenv
flask-cors
